Rename models/IP*.py to shorter, more Pythonic module names
